requirements.txt
yfinance
textblob
feedparser
jinja2
requests
html2image
tweepy==4.14.0
google-generativeai